derived data: Distinguish between dependent and predecessor derived data types
Summary:
Currently, it so happens that the only derived data type that has a predecessor is BSSM, and
its predecessor is skeleton_manifest which is also its sole dependent type.

This is not an inherent property, though.

For instance, when we introduce a new sharded type, we may want for the unsharded equivalent
to be its predecessor to allow fast backfilling from types that are already backfilled, but
marking the non-sharded version as a dependent type would cause it to always be derived when
we derive the sharded version, which would negate any deriviation speed benefit from sharding.

Reviewed By: YousefSalama

Differential Revision: D53766806

fbshipit-source-id: 04ae2f6df795efa3051685a8b034f88e4665ce83