StackActions: support "Edit stack" on partially obsoleted stack
Summary:
For stacks like:

  o
  |
  o
  |
  x
  |
  x

which happens if a stack is partially landed, or slighly more complex:

    x      o
    |      |
  o x    x o
  |/     |/
  o      x
  |      |
  o      x

The "o" (non-obsoleted) part is still linear and "Edit stack" can support them,
just select a subset of the hashes.

I think this also shows the strength of dag+set queries comparing to tree walks.

Reviewed By: evangrayk

Differential Revision: D53792685

fbshipit-source-id: 7eaf0b8c11bc8a813f3d93f5e959c648732c2dc4