Add edenfs_config and edenfs_upgrade logs to eden rage
Summary:
# Context
eden rage is a tool used to gather diagnostic output for debugging issues with eden. One of the things it does is collect a variety of logs and sends them to paste.

# This Diff
This diff adds 2 new log sources to eden rage,

1) edenfs_upgrade.log - written by a scheduled python script that restarts EdenFS on a weekly basis to ensure users are running with the most current release
2) edenfs_config.log - written by a scheduled python script that delivers EdenFS configs to user machines

# Technical Details
The logs are read on disk and sent to paste. The link to the paste is displayed in the rage output.

Reviewed By: genevievehelsel

Differential Revision: D53717899

fbshipit-source-id: d18545dfe224e6c4ad37688e75720159f1901400