Initialize persisted state from localStorage
Summary:
Previously, we used localStorage to store persistant data in vscode. In D53725181 we switched to use extension storage. But we need to add a migration path to read from localStorage the first time, so that we don't show the getting started UX after showing it already once before

This means when initializing the module-scoped object containing persisted storage values, if we see the extension-storage initialized data is empty, we try loading from localStorage. If we find any values in local storage, we send them to be persisted in the extension storage and clear localStorage.

Storage is accessed synchronously on startup, but we can't persist it back until we set up our communication bus. So we retry a few times until we can send the message properly. It's sad to have to do this, but this is a "one time" setup thing. We can delete this whole thing in a few weeks.

In my testing, it takes ~100ms to set up the message bus in development builds.

Reviewed By: quark-zju

Differential Revision: D53736603

fbshipit-source-id: 5e48ce588ef621b397c1478f478f339101442d0e