Persist temporary state
Summary:
Previously, we used localStorage in the webview to store state which persists across restarts.

This is unreliable. We've seen cases where localStorage just doesn't work. We can patch around that to avoid problems in our assumptions, but we need to have a new way to handle persisted state.

Our options for persisted state:
- localStorage (bad, as mentioned)
- acquireVsCodeApi().getState() / .setState() calls in the webview
    - This is supposed to be used for storing data that persists. But it only really works when the webview panel is remade after being hidden. Because we use `retainContextWhenHidden`, there's no need to do this. This state does NOT persist across closing and reopening the webview
- extension storage
    - This storage is persisted across restarts, but lives in the extension.

Our persisted state wants to be accessed synchronously. But since the storage lives in the extension, we need to be a bit tricky to make this work:
- load from persisted state in the extension just before creating the webview
- place the stored data into the webview HTML at webview creation time
- when initializing the platform, use this data as the initial value.
- write additional changes to the persisted data to the local webview object, but also send a message to the extension to persist the change any time it's updated
- for convenience, we send the entire local state to the extension when something changes, instead of sending just one key/value at a time

This approach allows us to get extension storage while having data accessed synchronously.

Note that this won't allow changes to the storage outside of ISL while ISL runs, but that's not a requirement we have.

Also, the step where we put the config directly into the HTML before loading the webview is a potential attack vector. We need to carefully validate the config. This technique is equivalent to `eval`-ing the user config. Only our extension should have access to this config, but it's still worth being cautious here.

TODO:
- we should temporarily also try to load from localStorage, so that we don't show the getting started UI an extra time
- we should have some debug tools in the UI to clear or read this storage state.

Reviewed By: quark-zju

Differential Revision: D53725181

fbshipit-source-id: 8b20840057f56b28b1004166a2795361c8c82bcf