Refactor cwd+cmd into ExecutionContext type
Summary:
Previously, it was common on the server to pass around 4 params together:
- cwd
- cmd (e.g. "sl")
- logger
- tracker

These things are needed to run commands in general (logger/tracker sort of tangential but in practice they're just about always passed as well)

This diff bundles these things together into one ExecutionContext type that we can pass around. This simplifies our APIs and makes it easier to other places to run commands outside of a Repository.

Reviewed By: quark-zju

Differential Revision: D53784113

fbshipit-source-id: 8e794fea8ddafea08bf957350ffc643c82b9288b