webview-app: prefer x64 chrome on Windows
Summary:
Update `find_chrome_like` to prefer x64 and chrome to x32 edge on Windows.
Windows ships with edge, if a user installs chrome it means the user probably
wants to use chrome instead of edge. If a user wants to use edge then they can
just ensure that chrome is not insalled.

Reviewed By: evangrayk

Differential Revision: D53829184

fbshipit-source-id: e659d4834b5667bd1f3e9c26893b126de12c44b9