tweakdefaults: always enable this extension
Summary:
This extensions is already enabled internally and in OSS - just not in tests. The biggest test fallout has been "log" requiring "--all" for previous behavior, and "rebase" requiring a destination.

Some commit hashes also changed due to the tweakdefaults "refresh date" behavior for "amend" and "rebase".

In particular, I want to make the "update" destination always required which tweakdefaults helps with since it makes "--dest" required for rebase.

This is one of those extensions that should be rolled into the core product. This is step one, in that it will require any incompatible tests to be udpated.

Reviewed By: zzl0

Differential Revision: D53622730

fbshipit-source-id: d53164312fbdd41afec8806c5b69b456cbe61490