merge: get rid of "linear" update check
Summary:
I think the only way this is used is if someone has commands.update.check=linear in their config. And even then it wouldn't take effect for HGPLAIN=1 since "commands" config items are dropped.

The code in updatetotally() looked like it defaulted to "linear", but the default in configitems.py is "noconflict", so that always won (even with HGPLAIN=1).

"linear" mode performed file merges when updating, but only if the source and destination commits were ancestors/descendants of each other. This could be useful in some situations - not sure. I'm trying to simplify the update checks, though, so let's get rid of it for now.

Reviewed By: zzl0

Differential Revision: D53622734

fbshipit-source-id: 989c76408d2852b1e06c6c0dca760cbcfae51809