checkout: use an enum to clarify/expand bookmark action
Summary:
Replace the inscrutable `Option<String>` with an enum to make the effect of the parameter more apparent.

And add a new bookmark action `None` that doesn't update the active bookmark at all. This will be used by Python to be equivalent to merge.goto() calls (which doesn't consider the active bookmark).

Reviewed By: zzl0

Differential Revision: D53622733

fbshipit-source-id: 619a4472a17449e3a7911ece334e8f94cf679f1f