s/hg/sl/ in editor filenames used by Sapling
Summary:
This ensures that when you run something like `sl commit`
or `sl histedit`, the filename for the temp file will end in
`.<action>.sl.txt`. In case you want to add special
syntax highlighting or LSP-like behavior, this makes it easier
to identify the purpose of the text file.

Reviewed By: muirdm

Differential Revision: D53818687

fbshipit-source-id: f994009fb116a0ac03c3d061ca14bb990764d886