Refactor TLSArgs argument for LFS Server and Metagit Location Service
Summary:
LFS Server relies on TLS args as one of its startup arguments if HTTPs is to be used. Because of previous limitations of clap, specifying an argument group that should either all be present or all be absent wasn't straightforward. That's why we defined `TLSArgs` as a clap argument containing all optional fields and manually validated in the caller if the all-or-none invariant is honored.

With the latest version of clap, we no longer have to go through this and can instead express it in much more ergonomic manner.

Reviewed By: gustavoavena

Differential Revision: D53771227

fbshipit-source-id: 6ec573d80b113b32d2984d125af53beebe5cae90