derived data: Reverse the `DerivationContext` <-> `DerivedDataManager` dependency
Summary:
**Before:**
* Each derived data type depends on `DerivationContext` which depends on `DerivedDataManager`

**After:**
* Each derived data type depends on `DerivationContext`
* `DerivedDataManager` depends on `DerivationContext`

**Why should I care?**
`DerivedDataManager` is the API used by mononoke to interact with data derivation.
Due to the fact that individual derived data types used to depend on `DerivedDataManager`, we
couldn't introduce the opposite dependency: from `DerivedDataManager` to individual derived data types.

This made it impossible to improve the ergonomics of the `DerivedDataManager` API, for instance, enabling the functionality to derive multiple derived data types at a time.

Reviewed By: mitrandir77

Differential Revision: D53811705

fbshipit-source-id: b4843511ab894cf2bc1ef3658ebdccb5022415dc