Open submodule dependency repos with mononoke app
Summary:
This diff starts removing the `TODO(T174902563)` comments to load the submodule deps from config.

This one handles specifically all the binaries that use Mononoke App API. This includes the forward syncer, which is the most important binary for this task.

The older ones using cmdlib/MononokeMatches are handled in the next diff.

Differential Revision: D53316969

fbshipit-source-id: ae6b0428441a1ab55b1c7caf2a7fa5571bfc92d0