Integration test validating actual round trippability
Summary: We realized recently that we are not validating round trippability because once the raw git commits are imported in Mononoke, we don't regenerate them from bonsais but instead use the original commits. As a result, we were not able to identify that the author name parsing from bonsai to git was broken. This integration test would help us identify regressions in the round-trippability in the future

Differential Revision: D53859498

fbshipit-source-id: 7ec0a5b72472601dad31358ab824468d84521a4a