Fix broken integration test
Summary: The test `test-newadmin-derived-data-list-manifests.t` broke as a result of landing D53921389 where I introduced tokio::spawn in two of the manifest ops methods. Due to the spawning of the futures in combination of buffered unordered, the result of list-manifest command became timing dependent. To avoid the test from failing in every alternate invocation, I have added `sort` at the end of the command to get predicatable output

Reviewed By: mitrandir77

Differential Revision: D53941711

fbshipit-source-id: 0d56f0f01edda9817a107c3c7361ad87962ecaf8