Update bookmarks get command to use is_tag instead of category
Summary:
The newadmin bookmarks get command looks at the category of the bookmark to determine if its a tag or not and based on that either displays the pointee commit or the pointee commit + tag metadata commit.

As we realized a while ago, the category column in bookmarks table isn't useful anymore and the full name of the bookmark is the actual source of truth. This diff updates the command to rely on the name instead of category. I will make such changes in other places in the code as well.

Reviewed By: singhsrb

Differential Revision: D53939926

fbshipit-source-id: 260f56b8174f8dca262557e085b40d980368efb2