Ignore bookmark category in gitimport
Summary:
Currently, gitimport is one of the few tools which respects category while creating a bookmark (setting is correctly based on the name of the bookmark). However, now that we want to get rid of category column completely it makes sense to start ignoring this column in the code completely.

Before landing this change I will update the bookmarks table to set the category to branch for all bookmarks.

Reviewed By: singhsrb

Differential Revision: D53941073

fbshipit-source-id: 3a77d129a2c8d14b5b861b962889350642eef488