Reduce single diff fetches
Summary:
We were slightly overfetching diffs in the sidebar. The intention was to have a debounced fetch for the single diff being viewed in the commit sidebar. So the sidebar commit will be up to date as quickly as possible.

Upon closer review, it seems like this was somewhat unnecessary:
- we want to fetch when the diff Id changes, so we have the latest info for the diff being viewed. This was working, but doesn't need any debouncing
- even with debouncing, it was triggering at least one additional unnecessary fetch
- the single diff info being refreshed is only really useful if the remote message changed, which we need to check on a timer anyway.
- since we fetch all diffs on a timer / focus change (including the single diff being viewed), the extra fetching during rerenders is like an extremely unreliable form of auto-refreshing one diff.

If we just make this a much simpler single fetch when the active diff changes, we reduce overfetching but things stay fairly up to date.

Reviewed By: quark-zju

Differential Revision: D53865096

fbshipit-source-id: 0df7c765cf602e71b8f0440a4e549bdbd90fd087