Show error notification if file open files
Summary:
If opening a file hits an error, it wouldn't show any error. We can forward vscode's error messages into the UI for a better experience.

Also, we were awaiting the promises in some cases, but this prevents processing other messages. Might as well use the `thenable` API instead of awaiting

Reviewed By: quark-zju

Differential Revision: D53873294

fbshipit-source-id: 677e14d2d4d720c6278087b9c73c2746f427a030