Don't log debug logging state all the time
Summary:
the onChange for this atom logged a message to the console. This shows up in every test, which is a little annoying.

Add a backdoor to `atomWithOnChange` to skip the effect on first mount. I could see that being sometimes useful, so it's more of an "only on change" if you need it to be.

Reviewed By: quark-zju

Differential Revision: D53864575

fbshipit-source-id: 8a89de99addf6bc03b8e69fee3897af970edd4af