Show configured file opener command in settings
Summary:
In D53866250 we added an sl config to customize how to open files from ISL. We don't allow you to directly set this from ISL for security, but we can show the configured value and offer a hint about how to configure it.

To do this properly, we need a small escape hatch in `configBackedAtom` to allow NOT json-parsing the config value. This lets us support non-quotes values that someone sets manually. We just manually JSON parse the string instead of doing it in the configBackedAtom handler.

Reviewed By: quark-zju

Differential Revision: D53869612

fbshipit-source-id: 0c68fbf14ccb51eaf31744485612079238a47324