warm_bookmarks_cache: log to seperate WBC scuba table
Summary:
WBC used to log to a separate scuba table but that was changed accidentally after migrating it to become a repo facet in D38349357. This diff fixes this so that it logs to a separate table again.
The table `mononoke_warm_bookmark_cache` was blocklisted as it was receiving no traffic for a long time. I'm too lazy to try to unblock it so I created a new table `mononoke_warm_bookmarks_cache`

Reviewed By: singhsrb

Differential Revision: D53952873

fbshipit-source-id: 1602c1a3cf7dcd3f638aacd660017737d489659a