Remove randomisation of bonsai_hg_mapping queries
Summary:
In D27821553, we added a random parameter to all *get* queries for `bonsai_hg_mapping` since this table gets queried multiple times to check if hg data has been derived for a commit or not. When this change was introduced, we did not have bookmark service so every WBC instance on every Mononoke / SCS task would end up querying this table for the same commit. This would lead to pylander killing the queries and hence this workaround was added.

With the introduction of Bookmark Service, we *_expect_* this to no longer be a problem. Only way to find out is to get ahead and test in prod ;)

Differential Revision: D53999740

fbshipit-source-id: 91ae88f3423fc71e815d41874fb18fc28586e137