derived data: Don't fetch what was just fetched
Summary:
I'm afraid I don't have an easy repro for this anymore, but this comes from a situation we observed with Youssef:

We were running a gitimport and were seeing that at the time of branch creation, multiple futures were competing to derive the same SkeletonManifest nodes, some of them failed at the `anyhow` line that is removed in this diff.

It appears that after fetching in the loop (`try_acquire_loop`), the next fetch would return `None` due to a race condition.

Side-step this race-condition by keeping track of the value we fetched instead of fetching it again expecting to find the same thing.

Reviewed By: YousefSalama

Differential Revision: D54003250

fbshipit-source-id: edc99c134a3f8b54071cb3053632997aa735cdfd