Add thrift type for BonsaiGitMappingCacheEntry
Summary: Similar to `bonsai_hg_mapping`, we want to add caching for `bonsai_git_mapping` since these mappings are static and do not change once they have been created. The first step to making the entries in `bonsai_git_mapping` table cacheable is to create a thrift type representing that entry to allow memcache and cachelib to cache this type.

Reviewed By: mitrandir77

Differential Revision: D54004262

fbshipit-source-id: f8a0127f9ff528b0876e9289ca2c16a3063379b6