CommitTreeList: remove tree based renderer
Summary:
Now the DAG renderer is at 100% and the only complaint seems to be
"not able to drag the avatar" which has an easy solution that is to
drag the title instead. Let's remove tree-based renderer so we can
focus improving the DAG renderer without worrying about compatibility.

This diff does not clean up CSS yet.

Reviewed By: evangrayk

Differential Revision: D54026595

fbshipit-source-id: 37f72512e02ee966e6b1d654288eb5f07e10743d