Add context for clippy warning
Summary:
In D53729268, we started using generic_asserts for better visibility in unit test failures. This makes us use the feature `generic_assert` but also forces us to use `core_intrinsics` which is not strictly needed. The only reason we are forced to use this is [this bug](https://github.com/rust-lang/rust/issues/120612).

Clippy doesn't like that we are using compiler internal stuff and gives a strong warning. Since we can't do anything about it, I have added a comment to provide context on why this cannot be fixed currently.

Reviewed By: YousefSalama, singhsrb

Differential Revision: D54061913

fbshipit-source-id: f5b952bfe13dc0e5af17b5b28c3d18adf43ba000