move merge_just_knobs earlier to de-flake the test
Summary:
merge_just_knobs (similarly to previous merge_tunables) unfortunately doesn't
wait for the config to be applied. By moving it to just beofore starting the
server (from just after) we can ensure that that config is applied when we
shoot the first request.

This problem was uncovered by D54030383 which removed the retries that papered
over the issue.

Reviewed By: YousefSalama

Differential Revision: D54063666

fbshipit-source-id: e4c54bb1bd375e8f44e9d3e65d9f98206145abf4