Add explicit return in switch
Summary:
This switch statement uses `raise_revlog_error()`, and then seemingly implicitly falls through to `case -3`. This was flagged as an unsafe fallthrough.

I'm not 100% clear if `raise_revlog_error` does something fancy with control flow such that it's actually a no-return. But we can avoid this confusino just by explicitly putting the fallthrough case underneath.

Reviewed By: quark-zju

Differential Revision: D54021725

fbshipit-source-id: b9aa66721f1cace51070b2a2f34586ec12d9a278