choose commit closer to mainline in case of ambiguity
Summary:
This is a fix for error exposed in previous commit. (At least in some cases).

After talking to markbt we arrived at conclusion that there are no truly bad choices here and in most cases it's better to choose something rather than error out. But in some cases it feels like the commit from large repo mainline is the more intuitive one. Let's say a single small commit gets synced twice: first time as a draft commit, second time during pushrebase as public. The more intuitive choice for further syncs of commit based on this one will be the public version.

This doesn't fix the issue completely but I feel it puts us in much better place.

Reviewed By: markbt

Differential Revision: D54012233

fbshipit-source-id: f4cef700c99bd49fdde261e053584942f840f591