consider only commits synced to mainline
Summary:
When we're syncing commits belonging to common_pushrebase_bookmark we need to
make sure that all the parents we not only synced but they were synced into
that branch. If they were synced to other branch we might need to sync them
again.

The code is not the prettiest here. The ancestor `hint` infra was created with
bias towards reducing the selection from many to exactly one and would always
return something if possible even if it didn't meet contraints and that's what
all the public functions did there. That's couldn't help with my stricter
requirements so I've created a new function there.

Reviewed By: gustavoavena

Differential Revision: D54012238

fbshipit-source-id: 72a15a39c32c90f3fab93e1e5eef49aaa1aa166c