move some utils into EdenTestCase
Summary:
I I wanna use some of this in an EdenHgTestCase, I can't inherit from this class
because EdenHgTestCase and EdenRepoTest are incompatible. I need to move these
methods down some of the layers, might as well put them on the widely used
EdenTestCase so that we can use them in all eden integration tests.

Reviewed By: genevievehelsel

Differential Revision: D50193866

fbshipit-source-id: da4bb8296b335939fec29f77d1abe5bf4e3ce76d