invalidate after a torn read so the file is accessible
Summary:
All future reads to these files will error until the file metadata caches
are cleared. This diff triggers a delayed invalidation after a file is
read with the incorrect size. This makes the file accessible again.
It's possible (though it doesn't always happen) that EdenFS's internal state
could be inocorrect for the file. a future `eden doctor` run will be able
to fix it up.

Reviewed By: MichaelCuevas

Differential Revision: D52278628

fbshipit-source-id: 580746bb91d21c79f87d16d5e442843617ac779c