oss: attempt to disable dependabot PRs
Summary:
The dependabot PRs for Javascript and Rust cannot be used as-is because our
monorepo requires "offline mirrors" for dependencies that dependabot cannot
directly update. So let's just limit the dependabot PR to 0. I guess this
does not disable security alerts which are shown separately
(https://github.com/facebook/sapling/security/dependabot).

Reviewed By: zzl0

Differential Revision: D54129988

fbshipit-source-id: b98cd903786b4a0245a17c0343733b22a1f000c6