Improve integration test before change
Summary:
Just improving the comparison of the `tree` output before and after the 2nd round of commits.
Now the vertical bars are properly aligned, which makes everything much easier to read.

I'm also moving the `hg show` command that displays which commit is checked out and the files it's changing to right above the `tree` comparison section.

This will clearly show that the deletion of submodules (i.e. entire submodule dependency) is working in D54110880.

Differential Revision: D54114660

fbshipit-source-id: 5106196bcde0b699b267cca0e06f4db36a60b87e