Handle deletion of submodules
Summary:
## This stack
For more details, see T174902563 or [this doc](https://fburl.com/gdoc/xik9okw8).

## This diff
Implement what I mentioned was a known gap in D54006211: properly deleting **entire submodules** in the monorepo version.

i.e. when you delete a submodule in the git repo, this will delete the file with type `GitSubmodule` in the Mononoke mirror, which in turn should **delete the entire expanded copy of the submodule** in its monorepo directory.

## How

- Check that the file being deleted is of type `GitSubmodule` in at least one of the parent commits.
- Get the entire working copy of the submodule commit that the source repo was pointing to.
- Generate `Deletion` file changes for all the files, pre-pending the submodule path to the files from the submodule.

Differential Revision: D54110880

fbshipit-source-id: 83265f11a1da7ed198f57e037d286d680ad96e16