checkout: normalize how we display conflicting files
Summary:
Now we show unknown file conflicts the same as normal conflicts. Previously we were inconsistent, depending on the type of unknown file conflict.

It might be nice to give specific information about unknown file vs tracked file conflicts, but either way the conflicting local changes show up in "status", so I think people will be able to figure it out.

Also add the "..and XXX more" line and the hint about how to deal with the conflicts.

Reviewed By: quark-zju

Differential Revision: D53882819

fbshipit-source-id: 1309c059ea9959e02b563f09ea4fd86e75c584dc