Fix loading persisted state in prod builds
Summary:
We were seeing the getting started UX appear on every ISL reload. This was working in dev mode, but apparently fails in production. That's because of the difference between prod builds and dev builds. Issues like this are very annoying.

In this case, in prod builds we were no longer adding the nonce to the webview content security policy (CSP), which is the only way the inline script tag that sets the previously persisted state could work. In dev builds, we have a relaxed CSP to get hot reloading to work.

I guess we could go back to non-hot-reloading dev builds for the vscode extension, so prod and dev match more closely. Maybe we should just use those kinds of builds for any changes that relate to the communication between webview and extension host or resource loading...? This issue was easy to miss among many other changes.

Reviewed By: sggutier

Differential Revision: D54213973

fbshipit-source-id: 85b19a1b1650de3d51c9fb2aee242d09f6e1d242