add a function for forcibly re-importing a particular ID from Mercurial Part 2
Summary:
# Context
When serving a request for a file or directory, EdenFS will check its own local caches before falling back to requesting Sapling, the Source Control client, for that file/directory object. When falling back to Sapling, it will check in the Sapling local cache before downloading the object from the remote server.

Sometimes the local cache file is corrupted or for other reasons we would like to force a pull from the remote server. This will be read back through the local cache to refresh it.

# This Diff
This Diff implements the get_remote_X functions proto'd in the previous diff.

# Technical Details
- ~~Implement get_remote_tree_iter, get_remote_content(_iter), get_remote_aux(_iter)~~
- update fetch and fetch_batch to have 2 booleans fetch_remote and fetch_local. Use the FetchMode to determine value of these bools.
- Add fetch_mode argument to get_content. Propagate this down to fetch and fetch_batch
- FetchMode enum is moved to types directory

# Discussion Points
~~- use of unimplemented!~~
~~- FetchMode is propagated through named functions eg. get_remote_tree->get_remote_tree_iter. This is due to the different namespaces(revisionstore, backingstore) having different and incompatble FetchModes. This could also be done by moving FetchMode to a shared location that both namespaces depend on.~~
~~- Contentstore is a legacy pathway used that according to the comments shouldn't be hit. I've chosen to ignore it for now and marked it as such with a comment.~~

Reviewed By: quark-zju

Differential Revision: D53975543

fbshipit-source-id: b8cb66c35b1fd19b0c7cf44692fa1faf87dd9acc