add a function for forcibly re-importing a particular ID from Mercurial
Summary:
# Context
When serving a request for a file or directory, EdenFS will check its own local caches before falling back to requesting Sapling, the Source Control client, for that file/directory object. When falling back to Sapling, it will check in the Sapling local cache before downloading the object from the remote server.

Sometimes the local cache file is corrupted or for other reasons we would like to force a pull from the remote server. This will be read back through the local cache to refresh it.

# This Diff
This Diff implements the c++ side skeleton for remote only file fetch. It does not implement any business logic. A followup PR will implement these functions.

# Technical Details
- Add new FetchMode RemoteOnly
- change localOnly bool to directly using FetchMode enum
- add skeleton for get_remote_XXX function

# Discussion Points

Reviewed By: genevievehelsel

Differential Revision: D53833131

fbshipit-source-id: 788459bca68d94fb1d14d997eb690619e454a371