add an eden debug command for forcibly re-importing a particular ID from Mercurial pt3
Summary:
# Context
When serving a request for a file or directory, EdenFS will check its own local caches before falling back to requesting Sapling, the Source Control client, for that file/directory object. When falling back to Sapling, it will check in the Sapling local cache before downloading the object from the remote server.

Sometimes the local cache file is corrupted or for other reasons we would like to force a pull from the remote server. This will be read back through the local cache to refresh it.

# This Diff
This Diff implements the debug CLI for FetchRemote. This is done by adding a call to getBlobRemote to the existing stub

# Technical Details
- Implement call to getBlobRemote
- Fix for FetchMode ffi

Reviewed By: genevievehelsel

Differential Revision: D54139848

fbshipit-source-id: 6aac3139d046c59de1f012732323b763afc902dd