tests: rework hgpython target
Summary:
For various reasons we'd like to rely on the exact same Python version we build our binary with (this will be used in a future diff). What's more, we also want the `sys.executable` name of the binary to be the actual binary we are running it with.

Luckily for us, `sl` already has a functionality where it behaves as Python if its renamed / executed as `python`, so let's use that instead of the wrapper script we had for running `hgpython`.

Reviewed By: muirdm

Differential Revision: D54087089

fbshipit-source-id: bafdbcaef8bb1d65ac3f789e39e3bafdf626b8a1