Eden: Make TOML parser a bit more user friendly
Summary:
When parsing a TOML file and there's an error, we're just raising an exception without providing the user useful information.

This diff will show the user the specific line that is likely to have the problem in the TOML file and for the cause of incorrectly quoted backslashes, a hint about the problem.

Reviewed By: jdelliot

Differential Revision: D54075769

fbshipit-source-id: 061b85341a2d834ad8aad8c8ea18d7af2491499c