DebugToolsMenu: add "force disconnect" button for debugging
Summary:
Sometimes it's nice to check the behavior if we lost connection to the
server (ex. disconnected wifi, or closing laptop lid). This button does
that. The duration can be changed via mouse wheel.

Not sure how this might work for the vscode message bus so it is only
implemented for the websocket version for now.

(Note: it seems the auto refresh stops working after a force disconnect..)

Reviewed By: evangrayk

Differential Revision: D54226608

fbshipit-source-id: 1f4131fb8b905d30bc4544b44e5ae59abfea218b