change renderObjectId/parseObjectId behavior
Summary:
# Background:

Rendering and parsing FilteredObjectIDs was previously broken. This is because we were naively stringifying all the bytes of the FilteredObjectIDs, some of which were whitespace/invalid when converted to utf-8. This made it impossible to use endpoints like `eden debug inode` and `eden debug blob` since the FOIDs weren't rendered/filtered correctly.

# This diff:

This diff changes the way we render/parse FOIDs so that they are rendered in a readable way, and they are copy/pastable for use in other debug endpoints. Eden will properly parse the copy/pasted FOIDs and convert them into FilteredObjectID structures in code.

Reviewed By: genevievehelsel

Differential Revision: D54150449

fbshipit-source-id: 9f8a4fcbf913bfd24a2c12a8220298ac3c856306