Support unified diff view
Summary:
Add an option to render the "SplitDiffView" component as "unified" diffs instead of only "split". Unified puts the removed lines and added lines one after another, isntead of side by side.

This is helpful for smaller width contexts, since you don't need to repeat unchanged context lines next to each other, so it's about half as wide.

For now, I use this for rendering suggestions in DiffComments. But I'll also add an option to change the rendering for the comparison view.

We could even use the width of the window to automatically switch unified/split.

We did have an existing unified views, used for chunk editing and interactive split. I opted not to reuse those, since this use case doesn't need chunk selections, and would rather have syntax highlighting etc, and it was hard to untangle the interactive split one.

We'll also need unified editors for the upcoming merge conflicts modal.

TODO:
- need some style tweaks to make borders behave as expected

Reviewed By: quark-zju

Differential Revision: D54275322

fbshipit-source-id: c34f009692910fa496b61d15c95902e70ba8a89f