tests: add a buck target for run-tests.py
Summary:
Very soon we intend to allow EdenFS to run in .t tests. In order to do this, we'll make EdenFS tests runnable with Buck only, so we need to add some code that imports part of the EdenFS CLI Python code / EdenFS integration tests. Now, we cannot simply import this EdenFS CLI code since it depends on thrift, which right now is only configured in the EdenFS codebase to be runnable with Buck.

So in order to achieve this, it's necessary to make `run-tests.py` a Buck target so that when it imports the EdenFS testing module it can properly do that. This diff is the first step for doing that.

Reviewed By: muirdm

Differential Revision: D54018306

fbshipit-source-id: 609599f7419845f6c37140180185c0d1b0850ce0