hg_mutation_store: fetch predecessors using recursive CTEs
Summary:
To fetch full predecessor chains without multiple hops to the database, we index entries by the "primordial" commit (i.e. the first commit in the chain).  However, for very long mutation chains, or mutations with a high divergence factor, this means we can end up fetching a very large number of entries.  We attempt to mitigate this by introducing a fetch limit, but this adds new problems where, for earlier commits in the chain, we can no longer fetch their predecessors fully.

Switch instead to using a recursive CTE query to walk the mutation history of the target changeset and fetch exactly the entries we need.

This change makes the record of the primordial commit, as well as the primordial index, no longer needed.  We continue to maintain it for compatibility during the upgrade and to allow us to roll back if necessary.   Later on, we can deprecate and remove them.

Reviewed By: liubov-dmitrieva

Differential Revision: D54304569

fbshipit-source-id: bb5fdec7482847832a783a36144119be6717675c